The TDS wire batch loop now continues after statement-terminating errors (severity 11-16 except deadlock; class ≥ 17 and NotSupportedException still abort), matching real SQL Server's default non-XACT_ABORT semantics — statement errors surface as a SimulatedErrorOutcome and the batch runs on, via a per-command continueOnError that only the wire opts into; sys.databases now projects the full 98-column SQL Server 2025 shape.

+++ SqlServerSimulator.Tests/CollationMetadataTests.cs
@@ -206,7 +206,7 @@ public sealed class CollationMetadataTests
         // the row-shape assertions to the simulated database by name.
         var sim = new Simulation();
         AreEqual("simulated", sim.ExecuteScalar("SELECT name FROM sys.databases WHERE name = 'simulated'"));
-        AreEqual((short)5, sim.ExecuteScalar("SELECT database_id FROM sys.databases WHERE name = 'simulated'"));
+        AreEqual(5, sim.ExecuteScalar("SELECT database_id FROM sys.databases WHERE name = 'simulated'"));
         AreEqual("ONLINE", sim.ExecuteScalar("SELECT state_desc FROM sys.databases WHERE name = 'simulated'"));
     }
